+++ source/gui/installerGui.py
@@ -309,6 +309,9 @@ def doSilentInstall(
 	startOnLogon = globalVars.appArgs.enableStartOnLogon
 	if startOnLogon is None:
 		startOnLogon = config.getStartOnLogonScreen() if not freshInstall else False
+	startAfterSignin = globalVars.appArgs.startAfterSignin
+	if startAfterSignin is not None:
+		config.setStartAfterLogon(startAfterSignin)
 	# Currently, this function is only called by ``core.main`` when ``--install`` or ``--install-silent`` are provided at the command line.
 	# The only use of the ``silent`` parameter to ``doInstall`` is to surpress the post-installation restart dialog.
 	# Since that dialog should be shown unless this genuinely is a silent installation, use presence of ``--install-silent`` as the actual value of the ``silent`` argument.